@@ -4,16 +4,16 @@ import 'package:excuses/ui/widgets/widgets.dart';
 import 'package:flutter/material.dart';
 
 class ExcusePageView extends StatelessWidget {
-  final List<Excuse> excuses;
-  final int currentPage;
+  final List<Excuse>? excuses;
+  final int? currentPage;
 
-  const ExcusePageView({Key key, this.excuses, this.currentPage})
+  const ExcusePageView({Key? key, this.excuses, this.currentPage})
       : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     var excuseCards =
-        excuses.map((excuse) => ExcuseCard.fromExcuse(excuse)).toList();
+        excuses!.map((excuse) => ExcuseCard.fromExcuse(excuse)).toList();
 
     return PageTransitionSwitcher(
       duration: Duration(milliseconds: 500),
@@ -25,7 +25,7 @@ class ExcusePageView extends StatelessWidget {
           transitionType: SharedAxisTransitionType.horizontal,
         );
       },
-      child: excuseCards[currentPage],
+      child: excuseCards[currentPage!],
     );
   }
 }
